Improved readability and ease of use
