+++ templates/about.html
@@ -18,11 +18,12 @@
 <br /><hr><br />
 <p><strong>Premise</strong></p>
 <p>
-    Cyber bullying has risen exponentially over the years, especially among teens. And while the traumatic experiences of the victims are well-known, little has been done by social media giants to preemptively take action. On a large scale application, merely acting on "reported posts" is not nearly sufficient. It is absolutely necessary to proactively participate in the prevention of cyber bullying.
+    Cyber bullying has risen exponentially over the years, especially among teens. And while the traumatic experiences of the victims are well-known, little has been done by social media giants to preemptively take action. On a large scale application, merely acting on the reported posts is not nearly sufficient. It is absolutely necessary to proactively participate in the prevention of cyber bullying.
 </p>
-<p><strong>About the Project</strong></p>
 <p>
-    <em>Mindhunters</em> is a state-of-the-art NLP-algorithm on which this social media platform is wrapped. It provides sophisticated detection of texts that are violent, offensive, sexist, racist, discriminant or derogatory in nature. The generated scores are used to provide alerts to the social media platform, which may take appropriate action against the post and/or user.
+    <strong>About the Project</strong></p>
+<p>
+    <em>Mindhunters</em> is a state-of-the-art LSTM-based NLP-algorithm on which this social media platform is wrapped. It provides sophisticated detection of texts that are violent, offensive, sexist, racist, discriminant or derogatory in nature. Scores are generated using Mindhunters, which affect the reputation of each user. The generated scores are used to provide alerts to the social media platform, which may take appropriate action against the post and/or user.
 </p>
 <p><strong>Credits</strong></p>
 <p>
